refactor: move `request` in to `shared/`
This removes all modules from the `common/` folder.
