--- packages/core/src/common/README.md
@@ -1,10 +1,0 @@
-# Common
-
-This folder is for code that can be used in any environment (eg: Node.js or Browser).
-
-### Examples
-
-Examples of what should **NOT** be in this folder:
-
--   `fs`: only works with Node.js
--   `window`: only exists in the Browser
