+++ packages/core/src/testE2E/amazonqGumby/transformByQ.test.ts
@@ -10,7 +10,7 @@ import * as CodeWhispererConstants from '../../codewhisperer/models/constants'
 import * as codeWhisperer from '../../codewhisperer/client/codewhisperer'
 import assert from 'assert'
 import { getSha256, uploadArtifactToS3, zipCode } from '../../codewhisperer/service/transformByQ/transformApiHandler'
-import request from '../../common/request'
+import request from '../../shared/request'
 import AdmZip from 'adm-zip'
 import { setValidConnection } from '../util/connection'
 import { transformByQState, ZipManifest } from '../../codewhisperer/models/model'
